pyproject.toml
[project]
name = "shrink-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.2.0",
    "einops>=0.8.0",
    "faiss-cpu==1.9.0",
    "ipdb>=0.13.13",
    "isort>=5.13.2",
    "ruff>=0.8.3",
    "sentence-transformers>=3.3.1",
    "torch>=2.5.1",
    "usearch==2.9.2",
]
